backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.11.5
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==5.0
python-multipart==0.0.18
python-dotenv==1.0.1
razorpay==1.4.1
email-validator==2.2.0
